attendance-system/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
pydantic==2.10.2
pydantic-settings==2.6.1
face-recognition==1.3.0
opencv-python==4.10.0.84
pillow==11.0.0
deepface==0.0.79
insightface==0.7.3
mediapipe==0.10.8
tensorflow==2.15.0
onnxruntime==1.16.3
scipy==1.11.4
numpy==1.24.3
geopy==2.4.1
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
cryptography==43.0.1